+++ app/models/user.py
@@ -7,5 +7,6 @@ class User(BaseIdUuid):
     __tablename__ = "users"
 
     name: Mapped[str] = mapped_column(String, index=True, unique=True)
-    
+    permission_level: Mapped[int] = mapped_column(default=1, nullable=False)
+    is_active: Mapped[bool] = mapped_column(default=True, nullable=False)
     spots: Mapped[List["Spot"]] = relationship("Spot", back_populates="created_by")