+++ app/schemas/user.py
@@ -13,8 +13,14 @@ class UserCreate(UserBase):
 class User(UserBase):
     id: int
     uuid: uuid.UUID
+    is_active: bool
+    permission_level: int
     
     model_config = ConfigDict(from_attributes=True)
+
+class UserUpdate(BaseModel):
+    is_active: bool | None = None
+    permission_level: int | None = None
 
 class UserSimple(BaseModel):
     id: int